Marshaler Refactor - Use Simple Attributes
Spiking on the Table and Object code, as well as other investigations
into possible marshaler functionality, indicates that using simple
attributes in the `Aws::DynamoDB::Client` class is the best
approach. Refactoring the marshalers to support this behavior.

Also includes another test of the attribute class to confirm use of the
database name attribute.