Update Default Table Name Logic
The previous table name assignment logic didn't account well for
modules. For example, the following class wouldn't generate a default
name that is valid in Amazon DynamoDB:

    MyModule::MyTable

You could still get a table to work with `#set_table_name`, but that's
not ideal. With this change, we will append modules to your table
name. For example:

    MyModule::MyTable => MyModule_MyTable