Refactor Attribute Tracking Code
This change introduces the `ModelAttributes` and `ItemData` classes for
tracking of model level attribute information, and item level attribute
data, respectively. These changes don't impact the user-facing API, but
should help improve understandability and maintainability of the
codebase.

There is one change to the external API, and that is that mutation
tracking is now on or off at the full table level, and cannot be toggled
on a per-attribute level.