Add Pagination Support Methods to ItemCollection
Includes unit tests, integration tests are pending. This change is
intended to make it more user-friendly to use the `#query` and `#scan`
methods with pagination exposed to outside callers. With the current
implementation, there is not a good way to stop and resume a query or
scan without shared state.