WIP: Add Save Support for Transact Writes
Will intelligently determine if a given item should be translated to a
"safe put" (a put that will fail if the item key exists on the remote
end) or an update operation. Force put is not supported, since that
should just be placed as a "put".

There is a new open question (added as a comment to a pending unit test)
about safe put operations, where they might smash any condition
expression the user provides. This needs to be handled in a sensible
way.