Add Error For Condition Expression With Safe Put
Intention would be to remove this error if possible after a notion of
expression building languages is added to aws-record. For now, there
isn't a reliable way to safely combine these two behavior patterns, so
I've added a...length...exception message in case this is ever hit.

Remaining work is down to documentation and refactoring, and possibly a
couple of additional integration tests.