Allow default value for String/Numeric Sets to be unset
The fix in #112 had the side effect of calling `type_cast` on `nil` if the `default_value` was unset.

Because the `StringSetMarshaler` and `NumericSetMarshaler` both convert `nil` to `Set.new`, this caused the unset default value in practice to become `Set.new` instead of `nil`.

Ultimately this caused downstream effects such as always persisting empty or unset sets as `nil` despite the `#persist_nil?` returning `false`.