Support custom update expressions when update has no other attribute changes
This is a minimal implementation to allow update expressions to be passed through directly to the api call. To avoid conflicts, an error is raised if an update expression would have been generated by the sdk to update dirty attributes.

Also: Adds support for pass-through API options to the class-level `.update` item operation.

Resolves aws/aws-sdk-ruby-record#137